+++ clock/time.go
@@ -3,6 +3,6 @@ package clock
 import "time"
 
 // DateTimeNow returns the current datetime in standard format
-func DateTimeNow() string {
+func DateTimeNow() DateTime {
 	return StandardFormat(time.Now())
 }