+++ my/request_test.go
@@ -5,4 +5,15 @@ import "testing"
 func TestRequest(t *testing.T) {
 	// TODO: NewRequest
 	// TODO: NewRequestAt
+	// TODO: Request.AddLog
+	// TODO: Request.AddFmtLog
+	// TODO: Request.AddDurationLog
+	// TODO: Request.AddErrorLog
+	// TODO: Request.SetNow
+	// TODO: Request.Output
+	// TODO: Request.SubRequest
+	// TODO: Request.MergeLogs
+	// TODO: Request.StartTransaction
+	// TODO: Request.CommitTransaction
+	// TODO: Request.AddTxStep
 }