+++ CHANGELOG.md
@@ -1,5 +1,7 @@
+## v0.3.2 - Model Package 
+  * Commit: 
 ## v0.3.1 - My Package 
-  * Commit:
+  * Commit: 2026-03-31 15:26
   * Request type
   * Request methods
   * Clock format functions